refactor: remove Kani proofs, update docs and module system
- Remove kani_helpers, protocol_proofs, rate_limiter_proofs, state_machine_proofs
- Remove mempool_proofs, chainstate_proofs, cryptographic_proofs, utxostore_proofs, rpc_proofs
- Remove get_proofs_by_tier.py
- Update module API, IPC, registry, integration, security
- Update network (ban_list, bandwidth, compact_blocks, dandelion, fibre, etc.)
- Update storage, RPC, payment, config
- Delete obsolete docs: ADDITIONAL_SECURITY_IMPROVEMENTS, PLAN_VALIDATION_REPORT, etc.

Made-with: Cursor

+++ src/network/compact_blocks.rs
@@ -69,7 +69,7 @@ pub fn calculate_tx_hash(tx: &Transaction) -> Hash {
         // Previous output hash (32 bytes)
         data.extend_from_slice(&input.prevout.hash);
         // Previous output index (4 bytes, little-endian)
-        data.extend_from_slice(&(input.prevout.index as u32).to_le_bytes());
+        data.extend_from_slice(&input.prevout.index.to_le_bytes());
         // Script length (varint)
         data.extend_from_slice(&encode_varint(input.script_sig.len() as u64));
         // Script
